Bug Fixes from Backlog, JJDD8HFV-19,21,22 and some code refactoring
