--- HolidaysJson2020.json
@@ -1,1 +1,0 @@
-{"holidays":[{"name":"New Year's Day","description":"New Year’s Day, on January 1, is the first day of the year and a public holiday in Poland.","date":{"date":1577836800000,"year":2020,"month":1,"day":1,"iso":"2020-01-01","dayInYear":1},"types":["National holiday"]},{"name":"Epiphany","description":"Epiphany, also known as the Procession of the Three Kings, is annually celebrated as a public holiday in Poland on January 6.","date":{"date":1578268800000,"year":2020,"month":1,"day":6,"iso":"2020-01-06","dayInYear":6},"types":["National holiday"]},{"name":"Valentine's Day","description":"Valentine’s Day (Walentynki) is one of the most romantic days of the year celebrated in many countries including Poland on February 14.","date":{"date":1581638400000,"year":2020,"month":2,"day":14,"iso":"2020-02-14","dayInYear":45},"types":["Observance"]},{"name":"March Equinox","description":null,"date":{"date":1584676176000,"year":2020,"month":3,"day":20,"iso":"2020-03-20","dayInYear":80},"types":["Season"]},{"name":"Good Friday","description":"Good Friday is observed in many churches in Poland to remember Jesus Christ’s death on the cross.","date":{"date":1586476800000,"year":2020,"month":4,"day":10,"iso":"2020-04-10","dayInYear":101},"types":["Observance"]},{"name":"Holy Saturday","description":"Holy Saturday, which is the day before Easter Sunday, is observed in many parts of Poland.","date":{"date":1586563200000,"year":2020,"month":4,"day":11,"iso":"2020-04-11","dayInYear":102},"types":["Observance"]},{"name":"Easter Sunday","description":"Easter Sunday is a public holiday in Poland to celebrate Jesus Christ’s resurrection.","date":{"date":1586649600000,"year":2020,"month":4,"day":12,"iso":"2020-04-12","dayInYear":103},"types":["National holiday"]},{"name":"Easter Monday","description":"Easter Monday, the day after Easter Sunday, is a public holiday in Poland. Both days celebrate Jesus Christ’s resurrection from the dead.","date":{"date":1586736000000,"year":2020,"month":4,"day":13,"iso":"2020-04-13","dayInYear":104},"types":["National holiday"]},{"name":"Labor Day / May Day","description":"Labor Day (Święto Pracy), also known as May 1 (1 Maja), is a national public holiday in Poland every year.","date":{"date":1588291200000,"year":2020,"month":5,"day":1,"iso":"2020-05-01","dayInYear":122},"types":["National holiday"]},{"name":"Day of the Flag","description":null,"date":{"date":1588377600000,"year":2020,"month":5,"day":2,"iso":"2020-05-02","dayInYear":123},"types":["Observance"]},{"name":"Constitution Day","description":"Constitution Day is an annual public holiday in Poland on May 3.","date":{"date":1588464000000,"year":2020,"month":5,"day":3,"iso":"2020-05-03","dayInYear":124},"types":["National holiday"]},{"name":"Mother's Day","description":"Mother’s Day is celebrated in Poland on May 26 each year.","date":{"date":1590451200000,"year":2020,"month":5,"day":26,"iso":"2020-05-26","dayInYear":147},"types":["Observance"]},{"name":"Whit Sunday","description":"Whit Sunday, also known as Pentecost or the Descent of the Holy Spirit (Zeslanie Ducha Swietego) is an annual public holiday in Poland 7 weeks after Easter Sunday.","date":{"date":1590883200000,"year":2020,"month":5,"day":31,"iso":"2020-05-31","dayInYear":152},"types":["National holiday"]},{"name":"Corpus Christi","description":"Corpus Christi is an annual public holiday in Poland on the Thursday of the ninth week after Easter.","date":{"date":1591833600000,"year":2020,"month":6,"day":11,"iso":"2020-06-11","dayInYear":163},"types":["National holiday"]},{"name":"June Solstice","description":null,"date":{"date":1592689420000,"year":2020,"month":6,"day":20,"iso":"2020-06-20","dayInYear":172},"types":["Season"]},{"name":"Father's Day","description":"Father’s Day (Dzień Ojca) is celebrated in Poland on June 23 every year.","date":{"date":1592870400000,"year":2020,"month":6,"day":23,"iso":"2020-06-23","dayInYear":175},"types":["Observance"]},{"name":"Assumption of Mary","description":"The Day of Assumption, also known as the Miracle upon the Vistula, is a national holiday in Poland on August 15 each year.","date":{"date":1597449600000,"year":2020,"month":8,"day":15,"iso":"2020-08-15","dayInYear":228},"types":["National holiday"]},{"name":"September Equinox","description":null,"date":{"date":1600781439000,"year":2020,"month":9,"day":22,"iso":"2020-09-22","dayInYear":266},"types":["Season"]},{"name":"All Saints' Day","description":"All Saints’ Day is a public holiday in Poland on November 1 each year.","date":{"date":1604188800000,"year":2020,"month":11,"day":1,"iso":"2020-11-01","dayInYear":306},"types":["National holiday"]},{"name":"Independence Day","description":"National Independence Day is an annual public holiday in Poland on November 11 to commemorate the country’s independence in 1918.","date":{"date":1605052800000,"year":2020,"month":11,"day":11,"iso":"2020-11-11","dayInYear":316},"types":["National holiday"]},{"name":"December Solstice","description":null,"date":{"date":1608544939000,"year":2020,"month":12,"day":21,"iso":"2020-12-21","dayInYear":356},"types":["Season"]},{"name":"Christmas Eve","description":"Christmas Eve is celebrated as a festive occasion in Poland on December 24.","date":{"date":1608768000000,"year":2020,"month":12,"day":24,"iso":"2020-12-24","dayInYear":359},"types":["Observance"]},{"name":"Christmas Day","description":"Christmas Day is a public holiday in Poland on December 25 each year.","date":{"date":1608854400000,"year":2020,"month":12,"day":25,"iso":"2020-12-25","dayInYear":360},"types":["National holiday"]},{"name":"Second Day of Christmas","description":"Boxing Day, also known as St Stephen’s Day or Holy Szczepan, is a public holiday in Poland on December 26 each year.","date":{"date":1608940800000,"year":2020,"month":12,"day":26,"iso":"2020-12-26","dayInYear":361},"types":["National holiday"]},{"name":"New Year's Eve","description":"Many people in Poland celebrate New Year’s Eve, which is the last day of the year in the Gregorian calendar, on December 31.","date":{"date":1609372800000,"year":2020,"month":12,"day":31,"iso":"2020-12-31","dayInYear":366},"types":["Observance"]}]}
