+++ pom.xml
@@ -18,7 +18,7 @@
             <plugin>
                 <artifactId>maven-assembly-plugin</artifactId>
                 <configuration>
-                    <finalName>CreateJar</finalName>
+                    <finalName>Holidays</finalName>
                     <appendAssemblyId>false</appendAssemblyId>
                     <archive>
                         <manifest>
