+++ src/main/java/com/infoshare/academy/highfive/holiday/HolidaysSingleton.java
@@ -1,4 +1,4 @@
-package com.infoshare.academy.highfive;
+package com.infoshare.academy.highfive.holiday;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -33,7 +33,7 @@ public class HolidaysSingleton {
         return this.holidaysList;
     }
 
-    public List<Holiday>  getHolidaysFilteredByName(String filter) {
+    public List<Holiday> getHolidaysFilteredByName(String filter) {
         validateHolidays();
         List<Holiday> filteredByName = new ArrayList<>();
         for (Holiday holiday : holidaysList) {
