+++ src/main/java/com/infoshare/academy/highfive/mapper/HolidayMapper.java
@@ -1,0 +1,15 @@
+package com.infoshare.academy.highfive.mapper;
+
+import com.infoshare.academy.highfive.holiday.Holiday;
+import com.infoshare.academy.highfive.view.HolidayDateView;
+import com.infoshare.academy.highfive.view.HolidayView;
+
+public class HolidayMapper {
+
+
+    public HolidayView mapEntityToView(Holiday holiday) {
+
+        return new HolidayView(holiday.getName(), holiday.getDescription(), new HolidayDateView(holiday), holiday.getTypes());
+    }
+
+}
