+++ src/main/java/com/infoshare/academy/highfive/menuconsola/HolidaysPlanningMenu.java
@@ -1,27 +1,29 @@
-package com.infoshare.academy.highfive.MenuConsola;
+package com.infoshare.academy.highfive.menuconsola;
 
+import com.infoshare.academy.highfive.holiday.HolidaysFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.InputMismatchException;
 import java.util.Scanner;
 
-public class HolidaysReviewMenu extends MainMenu {
+public class HolidaysPlanningMenu extends MainMenu {
     private static final Logger stdout = LoggerFactory.getLogger("CONSOLE_OUT");
 
     public static void runSubmenu() throws Exception {
 
-        HolidaysReviewMenu holidaysReviewMenu = new HolidaysReviewMenu();
-        holidaysReviewMenu.menuOptionsDisplay();
-        holidaysReviewMenu.getUserChoice();
+        HolidaysPlanningMenu holidaysPlanningMenu = new HolidaysPlanningMenu();
+        holidaysPlanningMenu.menuOptionsDisplay();
+        holidaysPlanningMenu.getUserChoice();
     }
 
     @Override
     public void menuOptionsDisplay() {
 
-        menuOptions.add("HOLIDAYS REVIEW");
-        menuOptions.add("Display employee vacation");
-        menuOptions.add("Display team vacations");
+        menuOptions.add("HOLIDAY PLANNING");
+        menuOptions.add("Search holiday by DATE");
+        menuOptions.add("Search holiday by NAME");
+        menuOptions.add("Add vacation");
+        menuOptions.add("Cancel vacation");
         menuOptions.add("Previous menu");
 
         stdout.info("\n\n" + menuOptions.get(0) + "\n\n");
@@ -30,7 +32,8 @@ public class HolidaysReviewMenu extends MainMenu {
 
             stdout.info(i + ": " + menuOptions.get(i) + "\n");
         }
-     }
+
+    }
 
     @Override
     public int getUserChoice() throws Exception {
@@ -53,10 +56,16 @@ public class HolidaysReviewMenu extends MainMenu {
         } else {
             switch (userChoice) {
                 case 1:
-                    stdout.info("\nDisplay employee vacation - UNDER CONSTRUCTION\n\n");
+                    HolidaysFilter.searchByDate();
                     break;
                 case 2:
-                    stdout.info("\nDisplay team vacations - UNDER CONSTRUCTION\n\n");
+                    HolidaysFilter.searchByName();
+                    break;
+                case 3:
+                    stdout.info("\nAdd vacation- UNDER CONSTRUCTION\n\n");
+                    break;
+                case 4:
+                    stdout.info("\nCancel vacation - UNDER CONSTRUCTION\n\n");
                     break;
                 default:
                     MainMenu.runMenu();
