+++ src/main/java/com/infoshare/academy/highfive/menuconsola/MainMenu.java
@@ -1,31 +1,38 @@
-package com.infoshare.academy.highfive.MenuConsola;
+package com.infoshare.academy.highfive.menuconsola;
 
-import com.infoshare.academy.highfive.HolidaysFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.InputMismatchException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Scanner;
 
-public class HolidaysPlanningMenu extends MainMenu {
+public class MainMenu {
     private static final Logger stdout = LoggerFactory.getLogger("CONSOLE_OUT");
+    public static int userChoice;
+    List<String> menuOptions = new ArrayList<>();
+    public String userChoiceString;
 
-    public static void runSubmenu() throws Exception {
 
-        HolidaysPlanningMenu holidaysPlanningMenu = new HolidaysPlanningMenu();
-        holidaysPlanningMenu.menuOptionsDisplay();
-        holidaysPlanningMenu.getUserChoice();
+    public static void runMenu() throws Exception {
+
+        MainMenu mainMenu = new MainMenu();
+
+        do {
+            mainMenu.menuOptionsDisplay();
+            userChoice = mainMenu.getUserChoice();
+        }
+        while (userChoice != 0);
     }
 
-    @Override
     public void menuOptionsDisplay() {
 
+        menuOptions.clear();
+        menuOptions.add("MAIN MENU");
         menuOptions.add("HOLIDAY PLANNING");
-        menuOptions.add("Search holiday by DATE");
-        menuOptions.add("Search holiday by NAME");
-        menuOptions.add("Add vacation");
-        menuOptions.add("Cancel vacation");
-        menuOptions.add("Previous menu");
+        menuOptions.add("HOLIDAYS REVIEW");
+        menuOptions.add("EMPLOYEES MANAGING");
+        menuOptions.add("CONFIGURATION");
 
         stdout.info("\n\n" + menuOptions.get(0) + "\n\n");
 
@@ -33,10 +40,9 @@ public class HolidaysPlanningMenu extends MainMenu {
 
             stdout.info(i + ": " + menuOptions.get(i) + "\n");
         }
-
+        stdout.info("\n" + "0: Exit" + "\n");
     }
 
-    @Override
     public int getUserChoice() throws Exception {
 
         boolean matchedToPattern;
@@ -51,25 +57,26 @@ public class HolidaysPlanningMenu extends MainMenu {
         }
         userChoice = Integer.parseInt(userChoiceString);
 
-        if (userChoice > menuOptions.size() - 1 || userChoice == 0) {
+        if (userChoice > menuOptions.size() - 1) {
             stdout.info("Wrong input - try again\n");
             getUserChoice();
         } else {
             switch (userChoice) {
+                case 0: System.exit(0);
                 case 1:
-                    HolidaysFilter.searchByDate();
+                    HolidaysPlanningMenu.runSubmenu();
                     break;
                 case 2:
-                    HolidaysFilter.searchByName();
+                    HolidaysReviewMenu.runSubmenu();
                     break;
                 case 3:
-                    stdout.info("\nAdd vacation- UNDER CONSTRUCTION\n\n");
+                    EmployeesManagingMenu.runSubmenu();
                     break;
                 case 4:
-                    stdout.info("\nCancel vacation - UNDER CONSTRUCTION\n\n");
+                    ConfigurationMenu.runSubmenu();
                     break;
                 default:
-                    MainMenu.runMenu();
+                    break;
             }
         }
         return userChoice;
