--- user.json
@@ -1,1 +1,0 @@
-{"holidays":[{"name":"New Year's Day","description":"New Year’s Day, on January 1, is the first day of the year and a public holiday in Poland.","date":{"date":1546300800000,"year":2019,"month":1,"day":1,"dayInYear":1,"iso":"2019-01-01"},"type":["National holiday"],"free":true},{"name":"Epiphany","description":"Epiphany, also known as the Procession of the Three Kings, is annually celebrated as a public holiday in Poland on January 6.","date":{"date":1546732800000,"year":2019,"month":1,"day":6,"dayInYear":6,"iso":"2019-01-06"},"type":["National holiday"],"free":true},{"name":"Valentine's Day","description":"Valentine’s Day (Walentynki) is one of the most romantic days of the year celebrated in many countries including Poland on February 14.","date":{"date":1550102400000,"year":2019,"month":2,"day":14,"dayInYear":45,"iso":"2019-02-14"},"type":["Observance"],"free":false},{"name":"March Equinox","description":null,"date":{"date":1553119112000,"year":2019,"month":3,"day":20,"dayInYear":79,"iso":"2019-03-20"},"type":["Season"],"free":false},{"name":"Good Friday","description":"Good Friday is observed in many churches in Poland to remember Jesus Christ’s death on the cross.","date":{"date":1555632000000,"year":2019,"month":4,"day":19,"dayInYear":109,"iso":"2019-04-19"},"type":["Observance"],"free":false},{"name":"Holy Saturday","description":"Holy Saturday, which is the day before Easter Sunday, is observed in many parts of Poland.","date":{"date":1555718400000,"year":2019,"month":4,"day":20,"dayInYear":110,"iso":"2019-04-20"},"type":["Observance"],"free":false},{"name":"Easter Sunday","description":"Easter Sunday is a public holiday in Poland to celebrate Jesus Christ’s resurrection.","date":{"date":1555804800000,"year":2019,"month":4,"day":21,"dayInYear":111,"iso":"2019-04-21"},"type":["National holiday"],"free":true},{"name":"Easter Monday","description":"Easter Monday, the day after Easter Sunday, is a public holiday in Poland. Both days celebrate Jesus Christ’s resurrection from the dead.","date":{"date":1555891200000,"year":2019,"month":4,"day":22,"dayInYear":112,"iso":"2019-04-22"},"type":["National holiday"],"free":true},{"name":"Labor Day / May Day","description":"Labor Day (Święto Pracy), also known as May 1 (1 Maja), is a national public holiday in Poland every year.","date":{"date":1556668800000,"year":2019,"month":5,"day":1,"dayInYear":121,"iso":"2019-05-01"},"type":["National holiday"],"free":true},{"name":"Day of the Flag","description":null,"date":{"date":1556755200000,"year":2019,"month":5,"day":2,"dayInYear":122,"iso":"2019-05-02"},"type":["Observance"],"free":false},{"name":"Constitution Day","description":"Constitution Day is an annual public holiday in Poland on May 3.","date":{"date":1556841600000,"year":2019,"month":5,"day":3,"dayInYear":123,"iso":"2019-05-03"},"type":["National holiday"],"free":true},{"name":"Mother's Day","description":"Mother’s Day is celebrated in Poland on May 26 each year.","date":{"date":1558828800000,"year":2019,"month":5,"day":26,"dayInYear":146,"iso":"2019-05-26"},"type":["Observance"],"free":false},{"name":"Whit Sunday","description":"Whit Sunday, also known as Pentecost or the Descent of the Holy Spirit (Zeslanie Ducha Swietego) is an annual public holiday in Poland 7 weeks after Easter Sunday.","date":{"date":1560038400000,"year":2019,"month":6,"day":9,"dayInYear":160,"iso":"2019-06-09"},"type":["National holiday"],"free":true},{"name":"Corpus Christi","description":"Corpus Christi is an annual public holiday in Poland on the Thursday of the ninth week after Easter.","date":{"date":1560988800000,"year":2019,"month":6,"day":20,"dayInYear":171,"iso":"2019-06-20"},"type":["National holiday"],"free":true},{"name":"June Solstice","description":null,"date":{"date":1561132454000,"year":2019,"month":6,"day":21,"dayInYear":172,"iso":"2019-06-21"},"type":["Season"],"free":false},{"name":"Father's Day","description":"Father’s Day (Dzień Ojca) is celebrated in Poland on June 23 every year.","date":{"date":1561248000000,"year":2019,"month":6,"day":23,"dayInYear":174,"iso":"2019-06-23"},"type":["Observance"],"free":false},{"name":"Assumption of Mary","description":"The Day of Assumption, also known as the Miracle upon the Vistula, is a national holiday in Poland on August 15 each year.","date":{"date":1565827200000,"year":2019,"month":8,"day":15,"dayInYear":227,"iso":"2019-08-15"},"type":["National holiday"],"free":true},{"name":"September Equinox","description":null,"date":{"date":1569225013000,"year":2019,"month":9,"day":23,"dayInYear":266,"iso":"2019-09-23"},"type":["Season"],"free":false},{"name":"All Saints' Day","description":"All Saints’ Day is a public holiday in Poland on November 1 each year.","date":{"date":1572566400000,"year":2019,"month":11,"day":1,"dayInYear":305,"iso":"2019-11-01"},"type":["National holiday"],"free":true},{"name":"Independence Day","description":"National Independence Day is an annual public holiday in Poland on November 11 to commemorate the country’s independence in 1918.","date":{"date":1573430400000,"year":2019,"month":11,"day":11,"dayInYear":315,"iso":"2019-11-11"},"type":["National holiday"],"free":true},{"name":"December Solstice","description":null,"date":{"date":1576988366000,"year":2019,"month":12,"day":22,"dayInYear":356,"iso":"2019-12-22"},"type":["Season"],"free":false},{"name":"Christmas Eve","description":"Christmas Eve is celebrated as a festive occasion in Poland on December 24.","date":{"date":1577145600000,"year":2019,"month":12,"day":24,"dayInYear":358,"iso":"2019-12-24"},"type":["Observance"],"free":false},{"name":"Christmas Day","description":"Christmas Day is a public holiday in Poland on December 25 each year.","date":{"date":1577232000000,"year":2019,"month":12,"day":25,"dayInYear":359,"iso":"2019-12-25"},"type":["National holiday"],"free":true},{"name":"Second Day of Christmas","description":"Boxing Day, also known as St Stephen’s Day or Holy Szczepan, is a public holiday in Poland on December 26 each year.","date":{"date":1577318400000,"year":2019,"month":12,"day":26,"dayInYear":360,"iso":"2019-12-26"},"type":["National holiday"],"free":true},{"name":"New Year's Eve","description":"Many people in Poland celebrate New Year’s Eve, which is the last day of the year in the Gregorian calendar, on December 31.","date":{"date":1577750400000,"year":2019,"month":12,"day":31,"dayInYear":365,"iso":"2019-12-31"},"type":["Observance"],"free":false}]}
